Move nav item component to it's own file
