+++ apps/full-site-editing/full-site-editing-plugin/wpcom-block-editor-nav-sidebar/src/components/nav-item/style.scss
@@ -1,0 +1,55 @@
+@import '~@wordpress/base-styles/colors';
+@import '~@wordpress/base-styles/variables';
+
+.wpcom-block-editor-nav-item {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	width: 100%;
+	height: auto;
+	margin: 0;
+	padding: ($panel-padding / 2) $panel-padding;
+	border-radius: $radius-block-ui;
+	text-align: initial;
+
+	&.is-selected {
+		background: $black;
+		color: $white;
+
+		.wpcom-block-editor-nav-item__label {
+			background: $dark-gray-500;
+		}
+	}
+
+	&:not( .is-selected ):hover {
+		background: $light-gray-400;
+	}
+
+	&:not( .is-selected ) .wpcom-block-editor-nav-item__slug {
+		color: $dark-gray-300;
+	}
+}
+
+.wpcom-block-editor-nav-item__title-container {
+	flex: 1;
+}
+
+.wpcom-block-editor-nav-item__title {
+	&.is-untitled {
+		font-style: italic;
+		color: $medium-gray-text;
+	}
+}
+
+.wpcom-block-editor-nav-item__slug {
+	font-size: $default-font-size - 2;
+	padding-top: 4px;
+}
+
+.wpcom-block-editor-nav-item__label {
+	display: inline-block;
+	padding: 4px;
+	border-radius: $radius-round-rectangle;
+	background: $light-gray-600;
+	font-weight: 600;
+}
