+++ apps/full-site-editing/full-site-editing-plugin/wpcom-block-editor-nav-sidebar/src/types.ts
@@ -1,0 +1,6 @@
+export interface Post {
+	id: number;
+	slug: string;
+	status: string;
+	title: { raw: string; rendered: string };
+}
